tests, remove example_config.toml
